Factor out commonalities back to PersistentConnectionProvider
